requirements.txt
Flask==0.7.2
beautifulsoup4==4.0.5
pymongo==2.1.1
requests==0.11.2
gunicorn==0.13.4
-e git+git://github.com/ohack/fwolin-python.git#egg=fwolin
